Split object from objects; update TODO
